+++ features/support/env.rb
@@ -9,16 +9,21 @@ Bundler.with_clean_env do
     gem_path = Dir['../../../tools/fastlane-plugin/fastlane-plugin-bugsnag-*.gem'].last
     `bundle config --local path vendor`
     `bundle install --gemfile=Gemfile`
-    `gem install #{gem_path} -i vendor`
+    `bundle exec gem install #{gem_path}`
   end
 end
 
-def fastlane_upload_symbols dsym_path
+def fastlane_upload_symbols(lane, dsym_path, api_key=nil, config_file=nil)
+  api_key_env = "BUGSNAG_API_KEY='#{api_key}'" unless api_key.nil?
+  config_file_env = "BUGSNAG_CONFIG_FILE='#{config_file}'" unless config_file.nil?
+
   Bundler.with_clean_env do
     Dir.chdir 'features/fixtures/fl-project' do
       `BUGSNAG_ENDPOINT='http://localhost:#{MOCK_API_PORT}'\
        BUGSNAG_DSYM_PATH='#{dsym_path}' \
-       bundle exec fastlane upload_symbols`
+       #{api_key_env} \
+       #{config_file_env} \
+       bundle exec fastlane #{lane}`
     end
   end
 end